update electron main

+++ tools/script/main/index.ts
@@ -480,13 +480,15 @@ Electron.ipcMain.on("message", (event: any, args: any) => {
 
 function InitUnpacked(){
     console.log("[main:init:unpacked]");
-    // Copy Net Files
+    // Copy Pack Files
     FileAPI.mkdirSync(path.join(__dirname, Electron.app.isPackaged ? "../../../../unpacked" : "../../unpacked"), {recursive: true});
     const net_files = FileAPI.readdirSync(path.join(__dirname, "../pack"), {withFileTypes: true});
     for (let item of net_files) {
-        let srcPath = path.join(path.join(__dirname, "../pack/"), item.name);
-        let destPath = path.join(path.join(__dirname, Electron.app.isPackaged ? "../../../../unpacked/" : "../../unpacked/"), item.name);
-        FileAPI.copyFileSync(srcPath, destPath);
+        if (item.name !== ".gitkeep") {
+            let srcPath = path.join(path.join(__dirname, "../pack/"), item.name);
+            let destPath = path.join(path.join(__dirname, Electron.app.isPackaged ? "../../../../unpacked/" : "../../unpacked/"), item.name);
+            FileAPI.copyFileSync(srcPath, destPath);
+        }
     }
     const child_process = require("child_process");
     if(os.platform() !== "win32"){
